move more cli logic into main package for external use
